# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShiftTrack API",
    "version": "0.1.0",
    "description": "Rutas derivadas de los routers implementados. Contrato de comandos generado de Zod. Los payloads de módulos se validan en sus handlers; este inventario todavía no describe todos sus cuerpos."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/media/upload": {
      "post": {
        "operationId": "post__media_upload",
        "summary": "POST /media/upload",
        "security": [
          {
            "deviceTransport": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/media/{id}/content": {
      "get": {
        "operationId": "get__media__id__content",
        "summary": "GET /media/{id}/content",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "firebaseIdToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token (no custom token)"
      },
      "deviceTransport": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Device TOKEN; solo transporte de pendientes y confirmaciones de equipo"
      }
    },
...
  }
}